wip: studydeletecontroller should
